+++ spec/integrations/any_fixture_spec.rb
@@ -6,6 +6,10 @@ describe "AnyFixture" do
   specify "it works" do
     output = run_rspec('any_fixture')
 
-    expect(output).to include("3 examples, 0 failures")
+    expect(output).to include("AnyFixture stats:")
+    expect(output).to match(/key\s+build time\s+hit count\s+saved time\n\n/)
+    expect(output).to match(/user\s+\d{2}\:\d{2}\.\d{3}\s+4\s+\d{2}\:\d{2}\.\d{3}/)
+    expect(output).to match(/post\s+\d{2}\:\d{2}\.\d{3}\s+1\s+\d{2}\:\d{2}\.\d{3}/)
+    expect(output).to include("4 examples, 0 failures")
   end
 end