+++ spec/integrations/fixtures/rspec/any_fixture_fixture.rb
@@ -5,14 +5,21 @@ require_relative "../../../support/ar_models"
 require_relative "../../../support/transactional_context"
 require "test_prof/recipes/rspec/any_fixture"
 
+# Detect Ruby 2.4+ by checking for Regexp#match?
+if /a/.respond_to?(:match?)
+  using TestProf::AnyFixture::DSL
+else
+  include TestProf::AnyFixture::DSL::Ext # rubocop:disable Style/MixinUsage
+end
+
 shared_context "user", user: true do
   before(:all) do
-    @user = TestProf::AnyFixture.register(:user) do
+    @user = fixture(:user) do
       TestProf::FactoryBot.create(:user)
     end
   end
 
-  let(:user) { User.find(@user.id) }
+  let(:user) { User.find(fixture(:user).id) }
 end
 
 describe "User", :user do
@@ -28,11 +35,21 @@ describe "User", :user do
   end
 end
 
-describe "Post", :user do
+describe "Post", :user, :transactional do
   let(:post) { TestProf::FactoryBot.create(:post, user: user) }
   after { post.destroy }
 
   it "creates post with the same user" do
     expect { post }.not_to change(User, :count)
+  end
+
+  context "with fixture post" do
+    before(:all) { fixture(:post) { TestProf::FactoryBot.create(:post) } }
+
+    let(:post) { fixture(:post) }
+
+    it "creates post with another user" do
+      expect(post.user).not_to eq user
+    end
   end
 end